requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
accelerate
qwen-vl-utils
Pillow
pyinstaller
psutil
pymupdf
tqdm
protobuf
sentencepiece
huggingface_hub
opencv-python
easyocr
mysql-connector-python


